requirements.txt
streamlit
pandas
torch
torchvision
pillow
numpy